Blockchain api tests
