+++ src/integration/java/api/BlockchainTest.java
@@ -1,0 +1,19 @@
+package api;
+
+import org.junit.jupiter.api.Test;
+
+import java.io.IOException;
+import java.util.Map;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasKey;
+
+public class BlockchainTest extends BaseClientTest {
+
+    @Test
+    void all() throws IOException {
+        Map<String, Object> actual = connection.api().blockchain.all();
+        assertThat(actual, hasKey("data"));
+    }
+
+}
